+++ rpg-graphic-engine/src/main/java/com/jukusoft/rpg/graphic/animation/Animable.java
@@ -1,0 +1,15 @@
+package com.jukusoft.rpg.graphic.animation;
+
+/**
+ * interface for objects which are animable
+ *
+ * Created by Justin on 11.12.2016.
+ */
+public interface Animable {
+
+    /**
+    * update frame
+    */
+    public void updateFrame (long now);
+
+}